Add files via upload
This is my second session with Svelte, and there are many things that make Svelte quite enigmatic. The seamless interaction between HTML and JavaScript is quite interesting. Today, I learned about running loops in Svelte. It's quite odd to see HTML perform loop functions, considering it is a markup language. I also learned how to create an inline function to delete data and display it on the webpage. Next, I will be learning conditional statements in Svelte